requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==72.1.0
wheel==0.43.0
